Switch all Pictures to use webp as fallback
With JPG we still have ~400 KB big images which are ~100 KB in avif. WebP is well established https://caniuse.com/webp and can be used as fallback of PNG and JPG.

For RS8 we had 100 MB dist folder before. 40 MB when using JPG+a few PNG. With WebP it is 33 MB.